Added if clause to not always try to remove a non-existent file.

+++ up_to_date_checker.sh
@@ -117,4 +117,6 @@ fi
 ## Cleanup
 rm "youtube-to-mp3_i386.deb"
 rm "youtube-to-mp3_x86_64.deb"
-rm "*.pkg.tar.xz"
+if [ -f "*.pkg.tar.xz" ]; then
+	rm "*.pkg.tar.xz"
+fi
